created modal file for Modal class
